normalized script names
