+++ explainaboard/metrics/meta_evaluation_test.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import unittest
 
-from explainaboard.metrics.nlg_meta_evaluation import (
+from explainaboard.metrics.meta_evaluation import (
     CorrelationWMTDAConfig,
     KtauCorrelationWMTDA,
     KtauCorrelationWMTDAConfig,
